requirements.txt
pyyaml>=6.0
click>=8.0
jinja2>=3.0
boto3>=1.26
checkov>=3.0
colorama>=0.4
tabulate>=0.9
jsonschema>=4.0
hcl2>=4.0
